cambios

+++ database/migrations/2021_01_16_001843_create_nota_tanque_table.php
@@ -22,7 +22,7 @@ class CreateNotaTanqueTable extends Migration
                 ->onDelete('cascade');
             $table->string('num_serie');
             //solo  salida
-            $table->integer('cantidad')->nullable();
+            $table->float('cantidad')->nullable();
             $table->string('unidad_medida')->nullable();
             $table->double('precio_unitario')->nullable();
             $table->enum('tapa_tanque', ['SI', 'NO']);
