+++ database/migrations/2021_11_15_143717_create_nota_talontanque_table.php
@@ -20,7 +20,7 @@ class CreateNotaTalontanqueTable extends Migration
                 ->on('nota_talon')
                 ->onDelete('cascade');
             $table->string('num_serie');
-            $table->integer('cantidad')->nullable();
+            $table->float('cantidad')->nullable();
             $table->string('unidad_medida')->nullable();
             $table->double('precio_unitario')->nullable();
             $table->enum('tapa_tanque', ['SI', 'NO'])->nullable();
